+++ textract/__init__.py
@@ -1,3 +1,3 @@
-VERSION = "0.5.0"
+VERSION = "0.5.1"
 
 from .parsers import process