+++ docs/changelog.rst
@@ -1,7 +1,7 @@
 Change Log
 ==========
 
-This project uses `semantic versioning <http://semver.org/>`__ to
+This project uses `semantic versioning <http://semver.org/>`_ to
 track version numbers, where backwards incompatible changes
 (highlighted in **bold**) bump the major version of the package.